feature userAccess : tests

+++ packages/types-database/src/types/provider/userAccess/ip/userIpV4.ts
@@ -1,9 +1,8 @@
 import { Schema } from "mongoose";
-import { Int32 } from "mongodb";
 import { type UserIpV4Mask, userIpV4MaskRecordSchema } from "./userIpV4Mask.js";
 
 interface UserIpV4 {
-	asNumeric: Int32;
+	asNumeric: bigint;
 	// for presentation only purposes
 	asString: string;
 	mask?: UserIpV4Mask;
@@ -11,7 +10,9 @@ interface UserIpV4 {
 
 const userIpV4RecordSchema = new Schema<UserIpV4>(
 	{
-		asNumeric: { type: Int32, required: true },
+		// Type choice note: Int32 can't store 10 digits of the numeric presentation of ipV4,
+		// so we use BigInt, which is supported by Mongoose and turned into Mongo's Long (Int64)
+		asNumeric: { type: BigInt, required: true },
 		asString: { type: String, required: true },
 		mask: {
 			type: userIpV4MaskRecordSchema,
